backend/requirements.txt
absl-py==2.3.1
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
curl_cffi==0.12.0
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
eventkit==1.0.3
executing==2.1.0
fake-useragent==2.2.0
Farama-Notifications==0.0.4
fastapi==0.121.2
filelock==3.18.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.57.0
frozendict==2.4.6
fsspec==2025.3.2
graphviz==0.21
greenlet==3.2.3
gymnasium==1.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.34.2
ib-insync==0.9.86
idna==3.10
imageio==2.37.0
immutabledict==4.2.2
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kagglehub==0.3.13
kiwisolver==1.4.8
Logbook==1.8.2
loguru==0.7.3
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
multitasking==0.0.12
narwhals==2.9.0
nest-asyncio==1.6.0
networkx==3.4.2
norgatedata==1.0.74
numpy==2.3.4
oauthlib==3.3.1
openai==2.8.1
ortools==9.14.6206
packaging==24.2
Panda3D==1.10.15
panda3d-gltf==1.2.1
panda3d-simplepbr==0.12.0
pandas==2.2.3
parso==0.8.4
peewee==3.18.2
pillow==11.1.0
platformdirs==4.3.6
playwright==1.54.0
plotly==6.3.1
prompt_toolkit==3.0.48
protobuf==6.31.1
psutil==6.1.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycparser==2.22
pydantic==2.12.4
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.1
pyotp==2.9.0
pyparsing==3.2.3
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
screeninfo==0.8.1
seaborn==0.13.2
setuptools==80.3.1
six==1.17.0
sniffio==1.3.1
snscrape==0.7.0.20230622
soupsieve==2.7
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.49.3
swig==4.3.1.post0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.9.0
torchviz==0.0.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.0
tweepy==4.16.0
twscrape==0.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
ursina==7.0.0
utils==1.0.2
uvicorn==0.38.0
watchfiles==1.1.1
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
win32_setctime==1.2.0
wrds==3.4.0
xgboost==3.1.1
yfinance==0.2.65
